% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{apply_filter_to_bitmap}
\alias{apply_filter_to_bitmap}
\title{Bitmap filter}
\usage{
apply_filter_to_bitmap(bitmap, filter_func)
}
\arguments{
\item{bitmap}{Bitmap created}

\item{filter_finc}{One of the filter functions}
}
\value{
Bitmap
}
\description{
Code to make the bitmap filter
}
