% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{binary_threshold_filter}
\alias{binary_threshold_filter}
\title{Binary filter}
\usage{
binary_threshold_filter(rgb_vec, cutoff = 127)
}
\arguments{
\item{rgb_vec}{Placeholder}

\item{cutoff}{Cutoff for binary filter}
}
\value{
Raw
}
\description{
Code to make binary filter
}
