% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupExamImageFilter-package.R
\docType{package}
\name{GroupExamImageFilter-package}
\alias{GroupExamImageFilter}
\alias{GroupExamImageFilter-package}
\title{GroupExamImageFilter: Filter for images}
\description{
Code as part of the group exam for filtering images
}
\author{
\strong{Maintainer}: Love Cyon \email{love.cyon@ki.se}

}
\keyword{internal}
