% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{Image_filter}
\alias{Image_filter}
\title{Image filter}
\arguments{
\item{image}{Filepath to image in ""}

\item{filter}{Name of filter; red, blue, green or magenta}
}
\value{
A filtered image
}
\description{
Code as part of the group exam for filtering images
}
\examples{
Image_filter("https://www.r-project.org/logo/Rlogo.png", green)
}
